Remove some deprecated calls and fix spelling

+++ src/main/java/hudson/plugins/global_build_stats/model/YAxisChartType.java
@@ -31,17 +31,17 @@ public enum YAxisChartType {
 
             double total = nbSuccess + nbFailures + nbUnstables + nbAborted + nbNotBuild;
             if (total == 0) {
-                dsb.add(Double.valueOf(0.0), Messages.Build_Results_Item_Legend_Statuses_SUCCESS(), range);
-                dsb.add(Double.valueOf(0.0), Messages.Build_Results_Item_Legend_Statuses_FAILURES(), range);
-                dsb.add(Double.valueOf(0.0), Messages.Build_Results_Item_Legend_Statuses_UNSTABLES(), range);
-                dsb.add(Double.valueOf(0.0), Messages.Build_Results_Item_Legend_Statuses_ABORTED(), range);
-                dsb.add(Double.valueOf(0.0), Messages.Build_Results_Item_Legend_Statuses_NOT_BUILD(), range);
+                dsb.add(0.0, Messages.Build_Results_Item_Legend_Statuses_SUCCESS(), range);
+                dsb.add(0.0, Messages.Build_Results_Item_Legend_Statuses_FAILURES(), range);
+                dsb.add(0.0, Messages.Build_Results_Item_Legend_Statuses_UNSTABLES(), range);
+                dsb.add(0.0, Messages.Build_Results_Item_Legend_Statuses_ABORTED(), range);
+                dsb.add(0.0, Messages.Build_Results_Item_Legend_Statuses_NOT_BUILD(), range);
             } else {
-                dsb.add(Double.valueOf(nbSuccess * 100.0 / total), Messages.Build_Results_Item_Legend_Statuses_SUCCESS(), range);
-                dsb.add(Double.valueOf(nbFailures * 100.0 / total), Messages.Build_Results_Item_Legend_Statuses_FAILURES(), range);
-                dsb.add(Double.valueOf(nbUnstables * 100.0 / total), Messages.Build_Results_Item_Legend_Statuses_UNSTABLES(), range);
-                dsb.add(Double.valueOf(nbAborted * 100.0 / total), Messages.Build_Results_Item_Legend_Statuses_ABORTED(), range);
-                dsb.add(Double.valueOf(nbNotBuild * 100.0 / total), Messages.Build_Results_Item_Legend_Statuses_NOT_BUILD(), range);
+                dsb.add(nbSuccess * 100.0 / total, Messages.Build_Results_Item_Legend_Statuses_SUCCESS(), range);
+                dsb.add(nbFailures * 100.0 / total, Messages.Build_Results_Item_Legend_Statuses_FAILURES(), range);
+                dsb.add(nbUnstables * 100.0 / total, Messages.Build_Results_Item_Legend_Statuses_UNSTABLES(), range);
+                dsb.add(nbAborted * 100.0 / total, Messages.Build_Results_Item_Legend_Statuses_ABORTED(), range);
+                dsb.add(nbNotBuild * 100.0 / total, Messages.Build_Results_Item_Legend_Statuses_NOT_BUILD(), range);
             }
         }
